use ArraysOfArrays to reduce allocation (#65)
* use ArraysOfArrays reduce allocation

* use ArraysOfArrays in doubly jagged too

* Let TLV also uses VectorsOfVectors

* Fix type instability by adding a new parameter for buffer

* Use map `ntoh` to eliminate BitVector

* Formatting

* Bump version, add compat

[skip ci]

+++ src/types.jl
@@ -46,8 +46,8 @@ end
     fDatime::UInt32
     fKeylen::Int16
     fCycle::Int16
-    fSeekKey::Integer
-    fSeekPdir::Integer
+    fSeekKey::Int64
+    fSeekPdir::Int64
     fClassName::String
     fName::String
     fTitle::String
